backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.12.0
anthropic==0.8.1
chromadb==0.4.24
langchain-text-splitters==0.0.1
pymupdf==1.23.21
python-docx==1.1.0
cryptography==42.0.2
numpy==1.26.3
httpx==0.26.0
tiktoken==0.5.2
pytesseract==0.3.10
Pillow==10.2.0
pandas==2.2.0
openpyxl==3.1.2
nltk==3.8.1
rouge-score==0.1.2
scikit-learn==1.4.0
